Add newScore function to new score file
